+++ src/views/resultsExplorer.ts
@@ -1,6 +1,6 @@
 'use strict';
 import { Functions, Strings } from '../system';
-import { commands, ConfigurationChangeEvent, ConfigurationTarget, Disposable, Event, EventEmitter, TreeDataProvider, TreeItem, window } from 'vscode';
+import { commands, ConfigurationChangeEvent, ConfigurationTarget, Disposable, Event, EventEmitter, TreeDataProvider, TreeItem, TreeView, window } from 'vscode';
 import { configuration, ExplorerFilesLayout, IExplorersConfig, IResultsExplorerConfig } from '../configuration';
 import { CommandContext, GlyphChars, setCommandContext, WorkspaceState } from '../constants';
 import { Container } from '../container';
@@ -8,7 +8,7 @@ import { RefreshNodeCommandArgs } from './explorerCommands';
 import { CommitResultsNode, CommitsResultsNode, ComparisonResultsNode, ExplorerNode, MessageNode, NamedRef, RefreshReason, ResourceType } from './explorerNodes';
 import { GitLog, GitLogCommit } from '../gitService';
 import { Logger } from '../logger';
-import { Messages } from '../messages';
+// import { Messages } from '../messages';
 
 export * from './explorerNodes';
 
@@ -16,6 +16,7 @@ export class ResultsExplorer extends Disposable implements TreeDataProvider<Expl
 
     private _disposable: Disposable | undefined;
     private _roots: ExplorerNode[] = [];
+    private _tree: TreeView<ExplorerNode> | undefined;
 
     private _onDidChangeTreeData = new EventEmitter<ExplorerNode>();
     public get onDidChangeTreeData(): Event<ExplorerNode> {
@@ -63,7 +64,8 @@ export class ResultsExplorer extends Disposable implements TreeDataProvider<Expl
         }
 
         if (initializing) {
-            this._disposable = window.registerTreeDataProvider('gitlens.resultsExplorer', this);
+            this._tree = window.createTreeView('gitlens.resultsExplorer', { treeDataProvider: this });
+            this._disposable = this._tree;
         }
     }
 
@@ -78,6 +80,10 @@ export class ResultsExplorer extends Disposable implements TreeDataProvider<Expl
     close() {
         this.clearResults();
         setCommandContext(CommandContext.ResultsExplorer, false);
+    }
+
+    getParent(element: ExplorerNode): ExplorerNode | undefined {
+        return undefined;
     }
 
     async getChildren(node?: ExplorerNode): Promise<ExplorerNode[]> {
@@ -125,14 +131,23 @@ export class ResultsExplorer extends Disposable implements TreeDataProvider<Expl
         this._onDidChangeTreeData.fire();
     }
 
+    async show() {
+        if (this._roots === undefined || this._roots.length === 0 || this._tree === undefined) return;
+
+        try {
+            await this._tree.reveal(this._roots[0], { select: false });
+        }
+        catch (ex) {
+            Logger.error(ex);
+        }
+    }
+
     showComparisonInResults(repoPath: string, ref1: string | NamedRef, ref2: string | NamedRef) {
-        this.addResults(new ComparisonResultsNode(repoPath, typeof ref1 === 'string' ? { ref: ref1 } : ref1, typeof ref2 === 'string' ? { ref: ref2 } : ref2, this));
-        this.showResults();
+        this.showResults(this.addResults(new ComparisonResultsNode(repoPath, typeof ref1 === 'string' ? { ref: ref1 } : ref1, typeof ref2 === 'string' ? { ref: ref2 } : ref2, this)));
     }
 
     showCommitInResults(commit: GitLogCommit) {
-        this.addResults(new CommitResultsNode(commit, this));
-        this.showResults();
+        this.showResults(this.addResults(new CommitResultsNode(commit, this)));
     }
 
     showCommitsInResults(results: GitLog, resultsLabel: string | { label: string, resultsType?: { singular: string, plural: string } }) {
@@ -160,18 +175,17 @@ export class ResultsExplorer extends Disposable implements TreeDataProvider<Expl
             return `${count === 0 ? 'No' : `${count}${truncated ? '+' : ''}`} ${resultsType.plural} for ${resultsLabel.label}${repository}`;
         };
 
-        this.addResults(new CommitsResultsNode(results.repoPath, labelFn, Functions.seeded(query, results), this, ResourceType.SearchResults));
-        this.showResults();
+        this.showResults(this.addResults(new CommitsResultsNode(results.repoPath, labelFn, Functions.seeded(query, results), this, ResourceType.SearchResults)));
     }
 
-    private async showResults() {
-        await commands.executeCommand('workbench.view.explorer');
-        Messages.showResultExplorerInfoMessage();
-        setCommandContext(CommandContext.ResultsExplorer, true);
+    private async showResults(results: ExplorerNode) {
+        await setCommandContext(CommandContext.ResultsExplorer, true);
+
+        setTimeout(() => this._tree!.reveal(results, { select: true }), 250);
     }
 
-    private addResults(results: ExplorerNode): boolean {
-        if (this._roots.includes(results)) return false;
+    private addResults(results: ExplorerNode): ExplorerNode {
+        if (this._roots.includes(results)) return results;
 
         if (this._roots.length > 0 && !this.keepResults) {
             this.clearResults();
@@ -179,7 +193,7 @@ export class ResultsExplorer extends Disposable implements TreeDataProvider<Expl
 
         this._roots.splice(0, 0, results);
         this.refreshNode(results);
-        return true;
+        return results;
     }
 
     private clearResults() {